fix: exception being raised when updating an existing island multiple times

+++ spigot/src/main/java/io/tofpu/speedbridge2/island/persistence/IslandEntity.java
@@ -5,7 +5,7 @@ public record IslandEntity(
         LocationEntity location
 ) {
     public record LocationEntity(
-            double x, double y, double z,
+            int x, int y, int z,
             float yaw, float pitch
     ) {
     }
